Add a wide thumbnail view option.
The default view limits the width of the thumbnail view.  This keeps us from
loading too many results at once, especially with small thumbs, which could cause
us to hammer the site a bit too much.  This option allows removing this limit.

To avoid loading too many results at once, this is a separate option from big and
small thumbs, and always uses big thumbs.  (If people modify the script to allow using
them together that's up to them, but let's be polite...)