Support the thumbnail data returned by some of the newer API calls.
/ajax/user/id/illusts/bookmarks returns full thumbnail data, but the
field names are all changed.  Switch to this naming convention (it's
probably what newer APIs will use), and remap keys from the older API
call (/rpc/illust_list.php) to this.

There isn't yet an actual replacement that I can find for /rpc/illust_list.php
that lets us just look up a list of thumbnail data for a list of IDs, so that's
still used for now.  The closest approximation is /ajax/user/user_id/profile/illusts,
but that's specific to a user.  One we can actually switch to this format we'll
be able to tell if a post is bookmarked in search results...