Show bookmark icons in the thumbnail view for bookmarked posts.
Finally found how to convince the site to return bookmark info, so add
bookmark icons to thumbnails.

This only shows whether an image is thumbnailed, it doesn't let you
edit bookmarks in the thumbnail view.

Bookmark icons are hidden when viewing your own bookmarks, since every
image on that page is bookmarked.

This attempts to show private bookmarks with the private bookmark icon,
but the site is a bit buggy and returns the wrong private flag on some
pages, so it doesn't always work.