Refactor to allow changing data sources without reloading.
Instead of loading a new page on navigation, we can now create a new data
source and continue in-place without navigating.  This makes navigating
quicker, and makes it so history navigation will just go back to the original
data source, so browser back won't reload the previous data.  That makes
quickly jumping to a similar images page and then returning to a search
work a lot better.

Some other changes in this:

- We do a better job at killing the underlying page in Chrome, which makes
things cleaner.  We also try to prevent the site from sending error logs,
so we don't spam the site's logs with errors caused by us if we can help it.
- The flash effect when returning from an image to thumbnails changed from
an outer glow to a brightness effect.  It looks a bit better.
- When returning to the thumbnail list, the image you were viewing is centered
in the thumbs if it would have been offscreen.  Previously, it would be brought
onscreen but only to the edge of the screen, making it harder to find.
- More data is prepopulated from search results, which makes some pages load faster.