Clean up loading pages that return tons of results in one call.
A few Pixiv pages return all of their results in one big lump: the illustration
list in the illust page, recommendations and related illustrations.  We handle
this reasonably well, but it still results in a hard to navigate scrollbar, and
a bit of load chunkiness.

Implement data_source_fake_pagination, which takes a big list of single-shot results
and paginates it locally.

This also makes data_source_current_illust not use data_source_from_page.
It does load from the page, but data_source_from_page is a separate thing
from data_source_fake_pagination, and we don't need most of the logic of
data_source_from_page, so we just bypass it.