Improve zoom support.
- Add a context menu for controlling zoom.
- Four zoom levels: 2x (the previous behavior), 4x, 8x, and fill.  Fill zoom scales the
image to cover the screen in both directions, which works well for a lot of images (but
doesn't do much for images that happen to fit the screen).
- Add zoom lock, which leaves zoom on, and click-dragging pans around without changing
the zoom.  Panning is in the same direction as it is when zoom panning.

Panning is locked to the image when in fill mode: the image won't pan past the edge, so
the screen is always filled.

The context menu opens with the zoom button under the cursor, so you can quickly toggle
zoom by right-left-clicking.