Rework thumbnail sizes with a thumbnail size slider.
The slider determines the number of pixels (w*h) we want each thumbnail to take,
and we figure out a thumbnail dimension that comes as close to that as we can,
taking the thumbnail aspect ratio and window size into account and allowing the number
of columns to be clamped, so we don't load too much data at once in searches.
This should generalize better to different window sizes and DPIs.