Allow zooming with the mouse wheel.
Either ctrl-mousewheel, or mousewheel while holding the right mouse button.

This zooms relative to the current display.  I haven't thought of a clean
way to display this, so there's nothing visible in the popup menu describing
this yet.

This is separate from the main zoom level and zooms in 50% steps.