Partial (not active) code for fixing the tag list when viewing an artist.
The problem with user tags right now is there are two completely different
APIs for showing an artist with and without a tag filter.  Showing without
a tag filter uses /ajax/user/id/profile/all, which returns a complete list
of IDs and no other data.  Showing with a tag filter is

/ajax/user/id/illusts/tag/tag_name?offset=0&limit=48

One returns all data and data is loaded page by page, and the other is
paginated at the API level.  This is a bit of a mess and hard to handle
(we want data_source_fake_pagination for big-lump APIs and a regular
data_source for paginated ones).