Switch most menus from hover to click.
Hover menus worked when the UI was smaller, but now there's too much "hover
over UI to display it, then wiggle the mouse around to get the random
dropdown menu to go away".  Switch most of them to click-to-show.  Better
polish would be to allow clicking the dropdown button and dragging to a
menu item, but for now it's just a toggle.

The avatar follow menu is still a dropdown.