Make the avatar context menu button work in the search page.
This allows you to tell if you're following the user of a search result by right
clicking.  This isn't displayed in the results directly since we have to make a
per-image request for it.

It also gives a quick way to get from search results directly to the artist search
for a search result.