Change the coordinate space of on_click_viewer.get_image_position and set_image_position.
Previously, get_image_position() returned normalized 0-1 coordinates and
set_image_position took those coordinates.  Change this to use 0-width,0-height
coordinates to make them more natural to work with and reduce the number of
coordinate systems involved.

Note that this is the scaled coordinate system and not the native resolution
of the image.  As long as we're always using on_click_viewer.width/height, this
doesn't complicate anything.