Add low-res image loading.
This uses the low-res thumbnail as a preview to show while the main image is loading.
It's blurry, but since they're often 1/50th the size of the full image, it gives us
something to display very quickly and doesn't slow down loading since it's so small.
It's the same image used as the search thumbnail, so it's often already loaded.
The main image loads on top of it, so it replaces the low-res preview as it loads.

The preview image is dimmed by default, to make it easier to tell whether you're seeing
the actual image.  This is similar to what the Pixiv iOS app does.  This can be disabled
in options (it might be jarring to have a dimmed image appear a lot).