Add an option to show the top UI on hover.
This is only really useful for people with very low-resolution monitors.

The UI shifts horizontally slightly between the two modes because CSS handles scrollbars
inconsistently between position: fixed and position: sticky.

Also, when in this mode clicks to the left and right of the UI will be eaten by the
UI box.