Fix white page on some pages if we're not logged in.
This was previously only handled on some pages and not others.  This also makes sure we
don't run helpers.block_network_requests if we're not active.

This doesn't stop us from doing everything: polyfills are still installed and some scripts
are still blocked.  This is hard to get right, since we need to install those things early,
but we don't know if we'll be able to run until the page is fully loaded, and there's no
way to run a script after page load and before page scripts run.