Implement tag translations and a quick search tag UI.
Tag translations are stored locally.  We won't have any to start, and we'll
fill them in as we see them.  There doesn't seem to be an API to batch look up
translations.  Not everything uses this for now.

The tag UI allows quickly toggling search terms on and off, using tags from
search history.

The search dropdowns can also now be resized horizontally.