Add support for Recommended Users.
Like the following page there's no special layout for this, we just show a few illustrations
from each recommendation, but it gives quick access to it.

The recommendations are based on a random sample of user IDs, probably randomly selected from
followed users.  This does what Pixiv's page does: takes that sample, and just keeps requesting
suggestions based on it.  This means that within a page, you'll generally get similar recommendations,
and you can keep scrolling for more similar recommendations.  If you don't like the suggestions,
hit refresh and it'll load a new set of seed user IDs, so you'll get new results.

This can also be used to view artists similar to another artist, by specifying a user ID explicitly
instead of using the random ones.

The display isn't perfect: it shows the user as a thumb (linking to the user), then the sample
illustrations (linking to the images).  It's a little weird since it won't line up with the
number of columns and the user entry tends to shift from spot to spot on each row.  It also uses
the user's profile image, which are too low resolution when the thumbs are large.  It's good
enough for now, though.