Refactor the remaining prefs a bit, so all prefs are added in menu_option.add_settings.
This way, all prefs are available in all prefs menus (except for thumbnail size).  Having
separate preference dropdowns with completely different options was weird, and it was easy
for someone to look at the options in the image view, only see a couple, and then never
bother looking at the identical-looking button in the search box to see the rest.