Allow holding shift while navigating to skip manga pages (#9, thanks amdost).
Home/end can be used to jump to the beginning or end to skip past pages, but
you need to do that manually when you hit something you want to skip.  Navigating
backwards jumps to the first page, so if you're skipping contents, you see the
title or main image instead of just the last page.