Support remembering and restoring the current page when browsing bookmarks.
Long ago this was handled by creating dummy thumbnail entries if we start
on a higher page, but that didn't work well.  If we started on page 100, it
would want to create thousands of empty placeholder.

That was removed since it didn't work well, but there's still a use for starting
on a later page.  If you have thousands of bookmarks, it's very hard to browse later
ones, since the only way to get to them is to scroll down page by page.  This isn't
important for most other searches, since they'll change too quickly for this to
matter, but it's useful for bookmarks.

Support this with a few pieces:

- A data source indicates that it wants this by returning true from supports_start_page.
- The first page to view is the page number in the URL when we start.  If it's supported,
the data source returns this from data_source.initial_page.  It's usually the "p" query
parameter.  We'll start loading pages from here.
- view_search will update the "p" parameter as the user scrolls down.  This doesn't create
a new data source, it just records where we left off if the page is reloaded.
- If we're not on the first page, a "load previous results" button is added to the top of
the results.  This just asks the data source to load page (data_source.initial_page - 1).
The data source also needs to update its initial_page to reflect this.
- view_search.refresh_images is updated to handle this more gracefully, inserting the new
results at the beginning instead of triggering a refresh and more intelligently setting
the scroll position.

Currently set_visible_thumbs and visible_thumbs_changed hardcode "p" being the page
number.