Rework disabling the context menu while the popup menu is open.
This is a hundred times harder than it be, since there's no way to know what element
was right-clicked in the first place during contextmenu and for some reason cancelling
mousedown doesn't prevent contextmenu, which it should since it's a result of the
default mousedown action.

THis fixes the context menu remaining blocked when it shouldn't be.