Rework zooming.
Previously:

- "Zoom lock" zoomed as if you were constantly holding down the mouse button
- "Zoom to full" set the base zoom to fill the screen ("0x" internally)
- "1x" meant to zoom the image to reach the edge of the screen without overflowing
- "2x" zoom meant 2x that base, "4x" meant 4x
- Mousewheel zooming was independent of all of this ("relative zoom")
- "Zoom to full" also disabled zoom lock if you were already zoomed to full (doing nothing
here was confusing)
- Panning was locked to the edge only when in "0x"

This needs a rework, I can barely keep it straight myself.  Merge "relative zoom" and the
zoom mode, and always make the zoom level relative to "unzoomed" (fill the screen without
overlapping).  Zooming to cover is just a specific zoom level instead of a separate mode,
and you can mousewheel-zoom between them directly.  This also adds an "actual" zoom level,
which is the image at its native size for viewing images that don't scale well (eg. screentones).
The overall behavior feels the same, but the different zoom states are a lot clearer.

Removed "2x" and "4x" buttons which aren't very useful, and added a button for "actual size".

We currently always lock to the screen edge.  Previously we only did this in some zoom
settings.  A setting can be added for this if needed, but let's try just enabling it for
a while and see if anybody complains.

Also, default to the top of the screen when viewing a portrait image instead of the center,
which works better for a lot of images (this is the same as the thumbnail panning).