Switch from XHR to fetch.
This rewrites the old GreaseMonkey workarounds to use unsafeWindow.fetch and
a bunch of workarounds for GM/FF nastiness.

I don't know if anyone is actually using GreaseMonkey and Firefox.  TamperMonkey
works a lot better in Firefox and GM is just a mess.  Don't expose Firefox security
wrappers to user scripts, that's not what they're for.