Rework zip_image_player to support incremental loading.
The original player supported it, but it did it by seeking around the
ZIP, making separate tiny requests for the central record and each
frame.  It was removed because it's really slow.

This does it with a single contiguous load, parsing files from the local
file headers instead of the central record.  This way, it's just as fast
since it's only making one request, but it starts playing as soon as it
has a frame available, so we get the best of both.

This supports seeking within downloaded video frames.