Store our scripts as JS strings too.
This lets us attach sourceURL to them, which gives us readable file and
line numbers in logs for release builds instead of showing "userscript:25000".
The URL will point to GitHub's raw file link, so it'll also load the original
source.  It isn't much different since we don't obfuscate code on build, but
it does allow showing it without string escaping.  The debug and release
builds are also more similar this way.