Simplify data_sources.current_illust.
Previously, this was treated as an artist search internally, since the page
contains a list of user posts.  This was a bit weird.  The data source for
two illusts would be the same if the user was the same, but there's no way
to know that without loading illust or thumbnail data for it to see who the
user is.

Instead, make data_sources.current_illust a dummy data source just to view
the image, and only use data_sources.artist for viewing an artist's posts.

This means data_source_class.get_canonical_url and page_manager.create_data_source_for_url
don't have to be async, which means we can always create a data source
without blocking.