Remove the auto-like setting, and just always auto-like on bookmark.
Let's just match the mobile app, and always like images when bookmarking
them.  I think this option is just bloating the settings dropdown and can't
think of any reason to want to disable it.  Liking an image without bookmarking
it makes sense, but bookmarking it without liking it just doesn't.