Change data-illust_id to data-id.
This was a bit messy: we had data-illust-id (hyphen) on the link, which was
always an illust ID, and data-illust_id (underscore) on the search result,
which could be an illust ID or a user ID.  Rename data-illust_id to data-id
since it can have different types of IDs.

Remove handling of these from get_illust_at_element.  These were getting
confused, it only needs to care about the data on the link (illustId/userId)
and not the search-specific data.