Fix the viewer not being destroyed immediately when exiting the view.
The call to stop_displaying_image was lost at some point, which causes
the viewer to be reused when exiting and reentering the view.  That wasn't
expected, and sometimes left the on_click_viewer not fully set up, and sometimes
not zoomed correctly.  This wasn't visible in normal use, but it can be
seen in quick view.