Simplify data_source_from_page.
This removes old logic for reading data from the original underlying
page to avoid fetches.  This used to be used for lots of things, but
almost everything has moved to real APIs now and this is only used for
a couple obscure pages, so the extra logic isn't worth the bother.