Try to fix some annoying default pan position behavior.
Previously this always kept the pan position when changing manga pages.  That's
fine when the images are the same aspect ratio, but strange when going from a
portrait image to landscape or vice versa, since the pan position wouldn't translate.

Try keeping the pan position only if the relative aspect ratio is staying the
same.  Going from landscape to landscape will stay in the same place, but going
from landscape to portrait will reset back to the top middle.