Tweaks to make the illust page less dependant on user info.
We load user info separately every time an image is viewed, but when you're
just viewing the image you never see any of it.  Remove the separate favicon
when viewing a bookmarked image, since it doesn't really serve any purpose
and it's the only place where we actually use user info.  This frees things up
to try to remove the extra load.