Refactor on_click_viewer.
Previously, the pan position was tracked as zoom_pos, which was the number of
pixels to offset the image from the top-left.  Refactor this to center_pos, which
is a normalized position within the image that should be centered.  At 0.5x0.5,
the image is centered onscreen.

This is useful because it's a value that transfers from one image to another, even
if they have different resolutions and aspect ratios.  This means that if we go
from a portrait image to another portrait image (which means we don't reset the
pan), we'll stay at the same position.  Previously if one image was much bigger than
the other, we'd end up panned to a weird spot.

This also handles screen resizing better, which also means it'll handle restoring
from history better if the window doesn't have the same dimensions as it did earlier.