Refactor creating thumbs.
Previously, we always added a page of thumbs, and expected the data source
to give a reasonable number of thumbs at once.  This works for Pixiv's
inherently paginated APIs, but it's inconvenient when viewing large local
directories.  Refactor screen_search to make it handle very large pages
better.

Thumbs are added incrementally and don't care about the page number.  If
we get thousands of images in a page, we'll only add a small number of them
at a time and add more as the user scrolls, so we don't freeze the UI and
try to add 10000 thumbs at once.

Handling coming back from illusts is also handled better.  If the user loads
a big search, clicks the first result, then mousewheel scrolls through iamges
for an hour, we can end up on an illust that's miles away from what was displayed
in the initial search.  Previously, every image that was viewed while in the
image view would be added to the search when you came back to it, since the data
source pages were all loaded and it just displays what the data source has.  It
can do the same thing now, but if there are way too many images to add as a result,
it'll discard results from the beginning and just add thumbs around the image you
landed on.  The earlier results are still accessible and are added if you scroll
up.