Fix the scroll position not being restored correctly when navigating from an image to a search if the search wasn't already loaded.
refresh_images was being called without waiting for the search to return data, so it would
only work if the data source was already loaded.  This made restoring the scroll position work
when exiting from an image back to a search, but not if you loaded directly into an image and
then navigated up to the artist/folder, since that search wasn't previously loaded.