Always center the last media ID instead of trying to figure out if it's onscreen.
This check doesn't work, probably because the node isn't visible and doesn't have
a scrollTop yet, which has been causing us to always center things that would be
offscreen when at the top of the results, and not anything right at the top.  Let's
just always center for now, since that's basically what's happening and the current
behavior is weird.