Make the context menu "out" button just return to the manga page view (and make it go to the new view).
Before, it would go from an image to manga pages for manga, image to the search for single
page posts, and from the manga display to the search.  This was pretty weird: you had to
press it twice if you wanted to go back from a manga page to the search.  It's hard to make
it work with the way the manga view URLs work (it doesn't retain the original search) and
there's a browser back button right there anyway, so let's make it only be used to jump to
the manga page view.  Use the back button (or browser back, or the browser back button on your
mouse) to go back to the search.