Change the way debug builds load.
Previously, debug builds loaded each script and resource with @require and @resource,
which were loaded by the bootstrapper.  This allowed refreshing the script to test without
having to somehow get the script into the script manager each time.

Switch this over to loading the script from a local server.  This is the same as the way
vview loads.  This has a few advantages.  It works in more script managers and browsers.
The other method only worked in TamperMonkey, only in Chrome and only if the user enables
an obscure option to allow file: URLs.  The script had to be updated whenever files were
added or removed, and the build script had to be run any time the stylesheet changed since
it couldn't generate it.  This method removes these limitations.