Try to work around problems with Pixiv site scripts.
Sometimes their scripts will run before we do and add event listeners on
popstate, and then navigate the page when we don't want it to.  Since it
registers listeners before we run we have no way of intercepting the event,
since they always receive it first.

Work around this by intentionally breaking Function.bind, so the code that
runs won't do anything.