Updated the UI for following users.
It's now a dropdown, so it's easier to see whether you're following publically or
privately, you can toggle it to private without having to unfollow first, and it
now fully supports editing follow tags.

There's no longer a button at the top level of the context menu that unfollows a
user.  It was too easy to unfollow a user when you just meant to click the avatar
to view the user.

The creepy eye icon is still in the avatar to show that you're following.