Fix up page handling in id_list.
id_list usually only contains the first page of manga posts (eg. illust_ids),
so if we're looking up a later manga page we need to check page 1.  Previously
this was handled by get_neighboring_media_id, but not by other callers to
get_page_for_illust, which was causing problems when navigating out of manga
posts.  Fix this by handling this in get_page_for_illust so it's done consistently.