Adjust the local image editing API.
Before, we cleared data like crops by setting it to an empty array, which
was a special case to set it to null.  Switch this to just setting the field
to null.  It's a bit annoying for an API field being null to be different from
it not being present, but the special case was too awkward.