Adjust the context menu dropdown.
Hide fewer menu options items when they're not available (grey them out
instead).  hide_if_unavailable makes sense for the download items because
there are a bunch of them and it clutters the list, but hiding things
in general makes it harder to find stuff.

Allow enabling slideshow for an image from a search, which navigates to
the image and starts the slideshow from there.