Adjust editing pan anchors.
Use shift to toggle editing the anchor instead of ctrl, so it doesn't conflict
with ctrl mode for the popup menu.

If shift is pressed, highlight the anchor button to show what it's doing.

Don't clamp the pan position while editing the anchor.  This makes it easier to
see what's happening when the pan is near the edge.