Fix the tag search dropdown in the search menu being cut off.
Making the search menu scrollable also hid the tag search dropdown.  This is
tricky to fix, and it's a bit weird to have two tag search fields right next
to each other anyway.  Chance this to a "Search Tags" menu item that just goes
to the search page.  This focuses the search field, so it doesn't take an additional
click to start a search.

This also makes some adjustments to the menu:

- Search Works is now at the top.  It's one of the more commonly used searches
and it was only at the bottom before because of the input widget.  Also, this
puts the mouse right over the tag search box after clicking it, instead of partway
down.
- Added icons.
- Renamed "New works" to "New works by everyone" to make it more consistent with
"New works by following".
- Moved "new works by following" above "new works by everyone", since it's much
more important.