Improve restoring the scroll position.
We were storing the first and last media ID that were nearby.  If we had thumbs
from pages 1 and 2 nearby before, the first could be on page 1 when we're starting
at page 2, so the first media ID wouldn't be loaded.  That prevented us from
restoring a similar set of nearby thumbs, so the scroll position wouldn't be restored.

Instead, store all media IDs that are nearby, so even if some of them are no longer
loaded, we can still find a similar match, which makes us restore the scroll position
more reliably.