Add a "refresh from start" button.
This replaces the "return to start" button that was only available at the
start of the list after reloading the page.

Having a separate button for this lets us support pages on the following search.
Most of the time I do want refreshing that page to start from the beginning, not
from the current page, so just enabling page support for that was inconvenient.
Having a separate button for this lets us handle both use cases.