Try to work around Chrome a content-visibility bug.
We watch for thumbs being near the screen using intersection observers, and
set content-visibility: hidden when they're far off.  In current versions of
Chrome, intersection observers occasionally seem to treat nodes with content-visibility:
hidden as not intersecting, which causes them to ping-pong between being marked visible
and not.  That isn't visible since it's happening offscreen, but causes the page to
chew CPU.

This seems like a Chrome bug, since content-visibility: hidden shouldn't affect intersection
observers.  A node that isn't being rendered still intersects the screen, and this is
needed to control content-visibility manually.

Try switching to content-visibility: hidden.  It seems to do well enough now.