Move the add and refresh bookmark tag buttons into the list.
Having them on the side is a pain for widget layout, since for some reason it's impossible
for a node to be overflow-x: visible and overflow-y: hidden at the same time.

This also removes the popup text for these for now.  It appears under the popup itself
and it's tricky to fix.