Replace helpers.State with an explicit list of open widgets that the user is busy with.
This generalizes stopping the slideshow while dialogs are open, which fixes the image
changing while the user is in dialogs like send image, etc.

This also pauses the slideshow in place, instead of letting it run to the end and then
pausing the image transition.  This makes what's happening clearer (you're not in a rush
to finish editing your bookmark before the image changes), and keeps it from sitting on
a black screen after the fadeout, which is ugly.