Try to fix some weirdness with widget visibility.
Setting the initial visibility is a bit weird.
The changes to viewer initialization caused quick view to stop handling mouse
movement correctly, because it made initialization non-async in a lot of cases.
That shouldn't matter.  The problem was that the widgets aren't actually visible
on construction until we return to the event loop, so they can't check things like
whether the mouse is over them.

Try to fix this: instead of deferring setting visibility, set it immediately and
only defer the initial visibility_changed and refresh, so widgets have their
initial visibility immediately.  This means that visibility_changed can sometimes
be called redundantly (it's simplest to just make sure this doesn't matter).