Work around PIL not being able to open some PNGs.
PIL refuses to open PNGs with truncated ICCP profiles.  That doesn't make sense:
most software doesn't use them, and other software doesn't do this, so this
causes PIL to be unable to open PNGs in the wild that everything else handles
fine.

ImageFile.LOAD_TRUNCATED_IMAGES can be used to disable this.  I'm guessing other
people ran into this problem and this flag was added as a hack to work around it,
since a real fix wouldn't be an undocumented global flag...