Refactor zoom factors
This changes zoom factors so a zoom factor of 1 is "cover" mode instead of "contain".
Zoom levels stay the same, converting in zoom_level_to_zoom_factor and zoom_factor_to_zoom_level.

This makes zooming and slideshows work the same way, so going back and forth is much simpler
and we don't have two completely different scaling modes to keep track of.