Update the way the script is distributed.
This changes the top-level script to @require the main script.  This avoids problems with
having such a large script: script manager editors can choke on them, and GreasyFork
has a 2MB limit, which we're slowly approaching.  We can do it this way now since GF
recently updated its policy to allow requiring scripts in this way if SRI is used.

This should also fix CSS source maps not working for other people, since the source
map will be stored along with the distribution.