Rework loading to try to reduce the effect of incompatibilities between script managers.
We don't actually need access to anything special from the script manager for almost anything.
The only thing we use is GM_xmlHttpRequest, and that's just for saving images to disk, since
Pixiv doesn't use CORS to give itself access to its own images.  Everything else can run as
a regular script in the page without anything extra at all.

Let's try loading in a different way: just create script nodes to load.  This is basically
the same as just loading unwrapped, since the script is just a script on the page without
anything extra.  We can give ourselves access to GM_xmlHttpRequest separately.

- All uses of unsafeWindow and unsafeDocument are now just window and document, since the
code isn't running in the sandbox.
- Removed all uses of exportFunction.  It's just a holdover from GreaseMonkey support (which
this might fix as a side-effect, though it's still not recommended).
- It's possible to run unwrapped now if image download support isn't wanted.
- The with(ppixiv) hack is still used, and the ppixiv object is always a global to allow this.
(That should be changed to something cleaner, but not here.)
- console.* are now pconsole.*, so we can hide site console logs without relying on unsafeWindow.