Make the delete button hover over the tags in the recent tag dropdown instead of reserving space.
These are only visible on the item that's being hovered, but they took up space in all of them,
which left a bunch of empty space in every row.  This makes better use of space, and while it
covers up the tag text a bit, it only does that while it's being hovered.