Added saved tags.
Before, tags were saved as recently searched whenever they were searched, pushing them to
the top each time.  This makes going back to tags easy, but it's a little awkard.  Searching
for a rarely-used tag near the bottom of your list doesn't mean it's suddenly commonly used,
so pushing it to the top can be annoying.  It's also awkward to just go down your saved tag
list one by one, since everything moves around as you search.

Add saved tags.  Tags can be saved from the recent list.  Once they're saved, you can
reorder them yourself, and they won't move around on their own when you use them.  It
acts more like a bookmark list of tags rather than a recent list that changes constantly.

Recent tags are still saved like before, so things will be the same for people who don't use
this.  Saved tags are listed at the top, so they won't be pushed down by having tons of
junk tags in recents.

This adds an editor UI for moving items around and saving tags.  Deleting items is now in this
UI instead of in the regular search interface.

The tag list at the top of the search is abbreviated unless it's hovered.  It's the search the
user just did, so he knows what it is, and this fixes big searches causing the entire header
to expand to fit it.