Change the way autocompletes are displayed.
Before, they were part of the dropdown.  This gets a bit confusing: if you enter text,
we either have to scroll up to it so it can be seen (which might not be wanted if you
just clicked on the search and we're trying to show the current selection), or not, in
which case you might not notice it's there at all, or have to scroll back up over and over
if we end up scrolled down slightly.

Change this to a separate list at the top that doesn't scroll, and the saved/recent tag
list below it scrolls.  This lets us keep the search list in a consistent place so the
autocomplete doesn't interfere with it, while also making it easy to use the autocomplete
results.