Use the aiType field, replacing our previous tag matching.
This doesn't work for rankings, but we'll fill that in manually.  It also won't work for
images loaded with the old /rpc/illust_list.php, but that's rarely used.

Unfortunately they didn't add any way to enable or disable filters for a specific search
or for recommendations, only as a global account setting.