Keep the search screen visible while viewing images.
Instead of hiding it, keep it visible and cover it with the image view.  This has the
same effect, but lets us animate the image view on and off on top of it, which isn't
possible if we hide the search.  It also slightly improves the browser back transition
on iOS, which always captures a black screen if we hide the search.