Fix briefly entering overflow-y: scroll, causing illust transitions to jump.
There was a brief period where we weren't on data-current-view="search" but
disable-scrolling wasn't set yet, so it fell back on the default of overflow-y: scroll,
causing the document to change width and thumbnails to move.  They went back
before becoming visible in the wrong position, but the transition would see
the wrong position.  Just always apply overflow-y: overlay regardless of screen,
so there's no gap.