Handle widget parenting better.
Search for the parent widget from the container.  This removes the need to pass in
the parent widget most of the time, and makes widgets know their parent much more
consistently, so shutting down is simpler and more consistent.  Parent widgets now
have a list of child widgets and the tree can be enumerated.