Add vv-container.
This just has a display: contents style.  This lets us have named containers that don't
affect layout, which is useful for places where we want an extra container node around
a node, but where adding a block around it would break layout, like when using flex to
center icons.  Use this to make the video UI layout more consistent.