Make the avatar on the user search smaller.
This used to poke out of the search box.  That wasn't intended and went away when the search
box was refactored, but now the avatar causes the search box to be a lot bigger than it needs
to be.  Just lower the avatar size a bit.  This also cleans up the avatar and follow dropdown
and makes the background on the dropdown consistent with other UI.