Improve expansion in the local tree.
If we expand a node automatically as the user navigates elsewhere, remember that the node was
automatically expanded, and collapse it automatically when we leave if the user hasn't navigated
in the tree explicitly.  This prevents huge chunks of the tree from becoming expanded if the
user navigates between directories in a bunch of different places during a search.