Remove sub-paths inside local searches.
Previously, if a folder was clicked inside a search, we'd remember the original search path
while displaying the folder.  This was so we'd still know the search we were in, which was
needed so the local navigation tree could be used to navigate the search.  This led to
confusing behavior: clicking a folder would still display the search as active, so it would
show "searching for bookmarks", but you'd be seeing files that weren't bookmarks, since you
were actually viewing the contents of a bookmark.

Remove this, since we don't do that with the local tree anymore, which should simplify
local navigation.