Simplify loading thumbnails.
Previously, we created a bunch of thumbnail entries, then as they became nearby (within a threshold
of the viewport), we'd request thumbnail data and set up the thumbnail if it was available.

These days thumbnail data is almost always available as part of results (we don't have to load it
separately), and this lazy loading is redundant with the get_media_ids_to_display logic.  Simplify
to always populate thumbs if possible.