Refactor get_media_ids_to_display a bit.
This had some weird interactions with the nearby IntersectionObserver.  Rework this so it
hopefully makes more sense: add more images in chunks as long as the last "nearby" thumb is
loaded.  Once we've added a thumb and it's not nearby, we don't need to add more until we
scroll down.