Rework dropdown popups.
Create menus in dropdowns dynamically when they're opened instead of statically.  This is easier
to work with.  It puts the popup at the document level instead of nested along with the button,
which makes more sense and makes it easier to position the dropdown and place it above other
UI.

Try to keep the popup from falling off the right edge of the screen.

Add an onvisibilitychanged callback, so we don't need MutationObservers to detect this.